requirements.txt
Pillow==8.2.0
sty==1.0.0rc1
